Remove prefixes from domain objects by default
There is a `use_long_name` option to allow for the full output, but this is not
something we'd normally want. Normally, the following should work:

```
.. js:class:: Foo()

    .. js:function:: bar()

        See :js:func:`baz`

    .. js:function:: baz()

        See :js:func:`Foo.bar`

See :js:func:`Foo.bar` or :js:func:`Foo.baz`
```

All these references should normally resolve. Because the javascript domain is
not like other domains, nesting does not work correctly for linking. Instead,
the following is currently required:

```
.. js:class:: Foo()

    .. js:function:: Foo.bar()

        This won't resolve :js:func:`baz`
        This will resolve :js:func:`Foo.baz`

    .. js:function:: Foo.baz()

        See :js:func:`Foo.bar`

See :js:func:`Foo.bar` or :js:func:`Foo.baz`
```

Unfortunately, all this results in extraneous prefixes to all the definitions
and a lot more typing. We can remove the extraneous prefixes at least, but
trimming the node after the javascript domain creates it.